Added files for working with resources
Resource in this context refers to anything that will be searched.
They are represented by objects which containes their text needed
to be searched. They receive input from classes in data_input but
themselfs dont extract any data. They are used as a way to transfer
data and make it easy to be searched.

They differ from data_input Classes in that data_input classes
extract data and convert to a form that can be searched while
themselfs uses requires output of data_input so that it can
easily be shared with other parts of the application like
search code(search engine).

This classes are very similar to input_data Classes but they
dont inherit each others methods. Resources classes act as a bridge
parts of the application. They keep data like title, type and text
of the resource. This data may be useful when searching or
diplaying results to the client.

Resources are also responsible for storing data so that it is
available for later use. They may use database, csv or some form
of text file to store the data which later can be used to create
resources objects.

For testing purposes, test/ folder is included for testing if Classes
bahave as expected.