Added C++ files for proccessing text data to be searched
Before a search can be performed, text needs to be proccessed to know what we are
working with. These files tries to implement a way of proccessing text to try to
understand what the text is. One could be knowing if characters are numbers or
not. They are included as part of cnavy search and most functionality is based
on features that will be used when searching.

These commited files fall in category of being character or characters. Characters
proccesss data about each individual character while characters is all about
multiple characters. Withing the files theres class called Characters and Character
that implemement the idea of character and characters. It is through these two CLasses
in which tokens and doc object(collection of tokens) will be build. Character Class may be
initialised millions of times same number as lenth of characters in text.

Performance is one of the things to be fixed on Character and Characters Classes. It was
taking more than 20s to create 5 million characters objects but it was improved to
less than 2s. 50 million characters takes about 10s from just close to 2m. Even if speed was
when creating the objects was improved. It takes more CPU and memory for that to be done,
optimisation for memory is needed to make text proccessing faster when searching.

All those may get fixed in comming commits.