Removed features on CHaracter Class that caused to use more memory and init time
Characters Classes were expected to hold evrything about themselfs but since
their object may be created millions of times, that may result in more memmory
used or more time creating the objects. It was done by removing any inheritance
it in which it was derived Class. Methods that wre not used were removed including
attributes they were working on.

Its base Classes(compute class) was their methods made static so that they can be accessed
by other Classes without object creating or inheriting their methods. Bool_Characters CLass
methods are now static instead of inhering Characters Class.

The idea is to try to reduse as much memory as possible at low level but making Classes staic
is way to there as they are not stored as part of any object.