Added optional 'truth_ratio' argument to Summary_Characters methods and docstrings updated
The extra argument would give full control on ratio of characters each method needs to true
for for method to return true. remeber that methods in Summary_Charactes Class exists in
Character Class. By default, bool_ratio is 1 which means all methods in corresponding
Characters objects needs to be true for the method to be true. Changing the value to less
than would return true if ratio of Characters objects that returns true is greater than
bool_ratio argument.

The Class is called Summary_Characters because its methods try to summarise all characters
Characters object.