Added bool_set.cpp and bool_set.hpp containing Bool_Set Class
Bool_Set Class is just a class to emulate dynamic BitSet of C++ Standard Library.
Since text processing part of the application uses collection of booleans, theres
a need for dynamic bitset which is not available is STL.The Bool_Set uses Vector
to store booleans and each operation on bools results in iteration.

All methgods are static so that they can be accessed without creating instance of Class.
Class has no constructor as alll of its methods are static.