Added static_compare_characters.cpp and static_compare_characters.hpp containing Static_Compare_Characters Class
Attempts to compare two Characters objects using Classes like Bool_Characters and related Classes.
Being able to compare Characters objects would allow to know if two strings are the same or simialar
in some way. This would help detect spell error in input text and avoid problems of not matching
words that just differs by one character.

The downside is that it may be slow to do that due to long iterations performed by Bool_Characters and
Ratio_Characters methods. Class not yet tested at runtime for errors but was able to compile meaning no
compile time errors. This is the hardest part to be coded and one of hardest part. It can be simplified by
creating Bool_Compare_Characters and Ratio_Compare_Characters Classes that comparison. It was hard to code
it and it may face performance issues due to lot of iterations.