Added some few methods to  Static_Sompare_Characters Class, fixed bugs on existing ones and added test file.
Class is now tested and works as expected as in the test file commited in this commit.
Remeber that the methods are static so they only depend on data from their arguments.
Comparison works great for charcters that are the same or similar without spell errors
or some ccharacters missing. e.g "names" and "name" may be compared with ratio of 0.8
but "surname" and "name" may match with ratio of 0.

The reason is that wrong characters are being compared. "s" was compared with "n" which
are not the same and "u" with "a" which also not the same. Everything will be false
which will result in ratio being 0. To fix the issue, characters should be mapped
first before being compared to avoid wrong characters from being compared. Changing
"surname" to "namesur" before comparing may help.

Its just about tricks to try to match two different words. The above may woek if there
are no spell errors. Imagine "name" was "nbme", its almost impossible to know if
thats similar to "surname". Trick is needed to try to change solve that but it does
not need to be slow or even dare take a second or half of it. Thousands of words
will be compared so perfomance matters. Unless C++ is faster as speed of light
the we should not worry about performance(speed).