Turned Characters_Token methods to constant methods with const keyword
Methods needs to be const since instances of this Class will be shared by multiple
Tokens and none of them are expected to edit the objects. Non cont methods cannot
be called on const instance. Constant methods were used because they can be called
on const object which is what will be done on Tokens.

Remeber instances of this Class interact with Characters object and files
that work on Characters objects. Instances this Class will most likey be
declared constants.