Added files for Reference_Token Class with one of files being containing tests
Reference_Token Class is create portions of Token that are not specific to
one Token but could be shared by multiple Tokens. It was created as result
of composition of Characters_Token Class. All of its functionality and methods
either references Characters_Token Class or call its methods.

Currently, the class contains all methods and attributes as Characters_Token
but differ in implementaion but results are all the same as it composite
it. Reference_Class references all methods of Characters_Token Class
in a way that changes to object of that class it referecnces changes
causes changes to be reflected on its members.

Its this Class in which Token Class will be created through inheritance. This Class
does not perform lot of calculatins but its attributes already have values which
would improve objects creation time. Since its attributes are just references, that
would improve memory usage and/or CPU usage since attributes are not copied but
referencing existing memory.

This theoretically seems to improve memory and CPU usage but results will be seen when
creating Document object.