Checked out changes from commit c878b4d83 for character.cpp and character.hpp and performed small edit for compatibility.
In the latest commit, everytime when performing action on Character object calculation is performed
using <cctype> library. Commit c878b4d83 keeps data about Character object in attributes and
its methods accesses or modify the values. Advantage is that Character object can be shared
by reference and everytime operatio is performed there wont be a need to calculate something again
as it was pre-computed.

Issue is wheather instances of this Class will be stored in Tokens or not. This class does not contain
index data which is something that is needed for a character in token. Maybe its reponsibility of Token
or Doc object. If not then new Class will need to be created which reference object of this Class. Another
Class will supply features specific to the Class instances. At the end of the there will be a Character
like object that references another Character object and also having independent members that can
change without impacting referenced characters object.