Added Shared_Characters Class source, header and test files
Shared_Character class is used to create instance that will contain Character objects
that are shared by multiple Characters objects. Think of it as central place were
Character objects get stored and everyone that needs certain Character object references
it instead of creating one construtor.. This would save thousants if not millions of Character
objects that could have been had to be manually created.

Object of this class can be created from string and charcters objects but anything
that contain data that can be used to create CHaracter objects could work if
implemented. For string, duplicate characters get removed and resultiong characters
are checked if already exists in stored map(char, Character obj). If character does
not already exists in map then Character object is created and then stored with character
being key and that CHaracter object being the value.

It was observed from task manager that memory was saved as only few MBs of RAM was used instead
of 1000 MBs when creating characters object from string. It did save a lot of memory which seems
to be promising and rewarding at the end. Problem is that it takes some time to remove duplicates
from string and it takes seconds to work with long strings with lot of this time spent checking
if element exists or removing duplicate characters from string. This was tested on 100 million
characters(same character) which to ok about 30 seconds but improved to 20 seconds.

This same thing should be applied to Tokens and Characters objects so that memory can be saved
when creating document object.