updated tests for Reference_Token Class
Reference_Token Class does not have any future waiting for it. It was intended to reference attributes
of CHaracters_Token in hope of saving memory through references. That was far from being true since
a reference uses memory of 4 bytes(same as int and pointers). Each reference used 4 bytes which is much
compared to most  attributes using 1 byte as they are mostly bools. It does not create any performance
benefits but degrates as it depends on another object. Its not worth to use the object both practically
and theoritically.

References and pointers will save memory if referenced object is large and that 4 bytes can be worth
compared to object of size megabytes. Instead of referencing individual attributes rather reference
the whole object. It was cause by my lack of practical knowledge of references and memory address
including how much memory occupied. Intensions was to optimise memory and performance so that multiple
documemets can be processed at the same time.

For now it will just be kept and its purpose will be replaced by the Class that it was referencing
it objects attributes. Characters_Token Class will take its place up until Tokens get created.