removed folder containing old source files for Line_Character Class
Line_Character Class is no longer used and is never referenced in the application. Its no longer
worth it as it depends on external contents to function. It was supposed to represent characters
of line or sentence which is hard to perform at characters level. It could be easily be replaced by
Tokens of sinle character.

Its features could be added to Tokens at a high level code other than low level in which data about
document is not known. Lot of changes may be expected as text processing code becomes complex
or more clear.