renamed folder 'text_preccessing' to  xtess
text_proccessing folder was intended to hold files that proccess text to be used when searching. The rename
is just to add falvour to the directories othe than just text_proccessing, search, resources or data_input.
Naming would allow feature to be able to be be identified as individual entity other than just as one. 'search' foder
may be named as spectal/magnal and resources named as manser or left as it is.

'x' represent unknown and tess for text proccessing. Both meaning using text proccessing to find hidden/
ubknown part in the text. 'c' in cnavy represent C/C++ languages and navy represent navigation. Both of
them mean iterating/searching or navigationg through text which is what is done when searching text.

Someone is just not good in naming things. In the meantime those may be  changed or not used at all and stick
with self explatory names like search.