Added CMake support
Manually creating object files and compiling is such a hardwork. Its easy when file has small
dependencies but becomes harder when dependencies are big as it will. Supporting CMake opens
opportunity to create software build with 2 command which is easy compared to having to maintain
paths to files before compiling. Running tests was such difficult but that will be simplified with
DocTest.

CMake is so fast as it does not recompile file that was not changed unlike compilers which
just compile any file even it was not changed. One advantage is that it takes care of everything including
compiler and linker flags which allows application to be compiled on multiple machines without having
to wory about paths and compiler installed.

Project structure was altered so much to try to accomodate CMake. Deleted files were mostly deleted
unintensionally for moving meaning that they lost their histories. But this files are just test files which
are will be rewriten for DocTest. New files were added some being empty header to help guide in creating
the application with other being CMakeList.txt files. Some folders were also renamed with data_input renamed to
crawler and search to angral and resources to rsm(Resources Manager).

CMakeLists.txt in project director is the main entry to create application. It includes CMakeLists.txt in apps folder
creates target for executable. The one in project directory add dependecy to executable target which is then
build to apps folder in iside /build.

The project structure may be modified to improve efficiency in developing the application.