Added test_proto executable target
test_proto for prototyoe testing allows to test how application will be have when finished. It
allows to test for crawler, rsm, etc combined into one instead of being isolated. It allows for integrating
testing if not wrong. But mainly its for testing features before adding them to final executable or
for playing with the features at local maching.

It wont be versioned as it will not be changed. Its meant for developer to play with the features and experiment
with what can be done. All changes will be kept to local repo without affecting remote repo. In other words
test_pro is not test executable but for playing with features integrated similar to how they would be in executable.

test_proto executable target links to same lib as cnavy executable target. its expected to have everything
from cnavy final features. It does not hoever link to testslib as changes to test files would result in it
rebuilding even if it does not depend on them.