Added shared objects for Character objects to Raw_Token
Closes #4

Raw_Token uses Characters object but Characters object does not use shared objects to store
common characters/letters. Characters object contain collection of Character object which
represent character. Raw_Token is expected to use shared objects for Character object even if
Characters object does not use one.

Adding support for shared objects for Characters objects could have solved the issue but would not free
dependecy of Raw_Token to Characters object. Shared object for Character objects was added to Raw_Token
as static attribute/variable. Being static variable would allow Character objects to be shared by multiple
objects of Raw_Tokens or instances of Classes that inherit it.

The issue is that Raw_Token depends on Character object to be initialised and setup its memeber variables.
The issue was solved by creating Characters object on fly inside constructor which is only used temporary.
New constructor to create Characters object from collection of Character objects was added which bypasses
the proccess of having to create character objects again. The Characters object is used to initialise
Raw_Token object but never get stored as it does not use shared objects. It looses its scope after
leaving the construtor which frees it.

It takes close to 4 seconds to create Raw_Token object of 200 000 characters. It depends on variation
of characters in text. Everything is being handled by shared object which ensures Character object
is only created ones and be shared with common characters. For English, different characters that can
be used in text are so small with about 24 Alphabets and just 128 ASCII character. Its expected to improve
CPU usage and memory when creating Raw_Tokens objects.

For 10 Million characters it never finished running even if it was only 'c' character repeated. The issue is
with creating the temporal Characters object. The other statements just finished in few seconds but creating
Characters object too minutes without completing. The issue may be with memory usage when creating the Characters
object as it does not use shared objects. Its more likely of being memory usage which slowed down my computer
and the linux environment.

Raw_Token object in real application will only be used with words which would just be 1-12 in most cases. Its
not worth it to make it handle millions. Millions of characters were just use to test how efficient it is when
creating Character objects.