Updated Raw_Token to use templates
Closes #5

It was painful and harder proccess to turn Raw_Token to use templates without affecting
the way it works. It uses Character object for making calculations provided by Bool_Characters
and Ratio_Characters. Changing type of character class would break everything especiallly if it
does not inherit Character Class.

The provided template argument is only applied to only portion but the rest is kept as Character object.
Characters object get created on fly from objects of template argument. Casting is first tried but if it
failes Characters object get created manually from text.

Templates and generic programming are so comfusing and harder to explain as type is not know. Maybe templates arent
supposed to be used in this sutuation but it seems only way to escape Character and Characters objects and replace
them with Basic_Character and Basic_Tokien.