Added libcurl and myhtml(C lang) libraries
Both the libraries could assist in crawling and extracting of text from webpage.
libcurl could be used for making requests and myhtml for extracting/parsing the returned
html data from libcurl.

myhtml was written in C meaning that considerations for C language should be taking into
account when communicationg with the library. libcurl on other hand is easy, simple and
was inspired by Python requests which also simple and straight forward.

Using the two libraries could free dependence on having to use command line 'curl' and
regex for parsing html. Both of the are unreliable with curl on command line not working
for most urls annd lacking ability to view data like status_code. Its hard to know what happened
by the program other than 'url retured empty string or failed'.

Regex for parsing html is failing. script tag cannot be matched, soetimes failes to also find
body tag. Using '.' causes stackoverflow which terminates the program which intensifies problems.
The regex is failing for realworld htl even if it passes the tested one which is oneline simple
html.

Implementing the libraries could resolve issues #1, #2 and #7 .